refactor: remove ViewModeSwitcher to Header
