+++ apps/next/src/components/Header/utils.ts
