--- apps/next/src/components/ViewModeSwitcher/index.ts
@@ -1,1 +1,0 @@
-export { ViewModeSwitcher } from './ViewModeSwitcher';
